Rename `json-encoding-modern.ts` -> `json-encoding-context.ts`.
