+++ packages/data-model/test/json-encoding-modern.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, it } from "@std/testing/bdd";
 import { expect } from "@std/expect";
-import { JsonEncodingContext } from "../json-encoding-modern.ts";
+import { JsonEncodingContext } from "../json-encoding-context.ts";
 import {
   DECONSTRUCT,
   FabricInstance,
